+++ project/package.json
@@ -6,6 +6,7 @@
     "@reduxjs/toolkit": "^1.5.0",
     "axios": "^0.21.1",
     "classnames": "^2.3.1",
+    "dayjs": "^1.11.4",
     "history": "^5.2.0",
     "leaflet": "1.7.1",
     "react": "^18.2.0",
