+++ project/src/components/locations-item/locations-item.tsx
@@ -2,7 +2,7 @@ type LocationsItemProps = {
   state: string
 }
 
-export function LocationsItem({ state}:LocationsItemProps): JSX.Element {
+export function LocationsItem({ state }:LocationsItemProps): JSX.Element {
   return (
     <li className="locations__item">
       <a className="locations__item-link tabs__item" href="/">
