+++ project/src/components/offer-list/offer-list.tsx
@@ -1,16 +1,32 @@
+import { useState } from 'react';
 import { Offer } from '../../types/offer';
+import { CardOffer } from '../card-offer/card-offer';
 
 type OfferListProps = {
   offers: Offer[],
-  nearbyOffer?: boolean,
-  renderOffer: (offer:Offer, nearby?: boolean, key?: string) => [],
+  nearbyOffer?: boolean
 }
 
-export function OfferList({offers, renderOffer, nearbyOffer}: OfferListProps): JSX.Element {
+export function OfferList({offers, nearbyOffer}: OfferListProps): JSX.Element {
+  const [activeOffer, setActiveOffer] = useState({});
+
+  const changeActiveOfferHandler = (offer:Offer) => {
+    setActiveOffer({...activeOffer, ...offer});
+  };
 
   return (
     <>
-      {offers.map((offer:Offer) => renderOffer(offer, nearbyOffer))}
+      {
+        offers.map(
+          (offer:Offer) => (
+            <CardOffer
+              key={offer.id}
+              offer={offer}
+              changeActiveOfferHandler={changeActiveOfferHandler}
+              nearbyOffer={nearbyOffer}
+            />
+          ))
+      }
     </>
   );
 }
