+++ project/src/components/reviews-list/reviews-list.tsx
@@ -11,11 +11,11 @@ export function ReviewsList({reviews}: ReviewsListProps): JSX.Element {
     <section className="property__reviews reviews">
       <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviews.length}</span></h2>
       <ul className="reviews__list">
-
-        {reviews.map((review) => (
-          <ReviewsItem key={review.id} review={review}/>
-        ))}
-
+        {
+          reviews.map((review) => (
+            <ReviewsItem key={review.id} review={review}/>
+          ))
+        }
       </ul>
       <CommentSubmitForm />
     </section>
